requirements.txt
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
google-api-core==2.10.2
google-auth==2.14.0
google-cloud-core==2.3.2
google-cloud-pubsub==2.13.10
google-cloud-storage==2.5.0
google-cloud-vision==3.1.4
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.56.4
grpc-google-iam-v1==0.12.4
grpcio==1.50.0
grpcio-status==1.50.0
idna==3.4
numpy==1.23.4
opencv-python==4.6.0.66
pandas==1.5.1
Pillow==9.3.0
proto-plus==1.22.1
protobuf==4.21.9
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.2
pytz==2022.6
requests==2.28.1
rsa==4.9
six==1.16.0
urllib3==1.26.12
